refactor : price 자료형 : Long -> BigDecimal

+++ src/main/java/store/myproject/onlineshop/domain/account/dto/AccountDepositRequest.java
@@ -3,6 +3,8 @@ package store.myproject.onlineshop.domain.account.dto;
 import jakarta.validation.constraints.NotNull;
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 @AllArgsConstructor
@@ -10,5 +12,5 @@ import lombok.*;
 public class AccountDepositRequest {
 
     @NotNull
-    private Long depositPrice;
+    private BigDecimal depositPrice;
 }
