+++ src/main/java/store/myproject/onlineshop/domain/account/dto/AccountWithdrawRequest.java
@@ -2,11 +2,13 @@ package store.myproject.onlineshop.domain.account.dto;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class AccountWithdrawRequest {
 
-    private Long withdrawPrice;
+    private BigDecimal withdrawPrice;
 }
